rgw/role creation and listing: take user (owner) into account
This is the follow-up change to the previous commit which introduced the
"owner" field for the RGWRole object. This commit introduces the
functionality how this field would be populated other than the default
value (i.e., empty string).

rgw admin commands (role create & list) can now utilize "--uid" option
so that an admin can create a role and list available roles for a
specific user.

In addition, when a rest user/client creates a role, it'll be recorded
as the owner of the role.

Signed-off-by: Oguzhan Ozmen <oozmen@bloomberg.net>