python-common: fix osdspec_affinity check
When no `service_id` is provided to service spec (osd) it results in
OSDs created with "osdspec_affinity" attribute set to a string
containing "None".

The DriveSelection class relies on the comparison of the actual
value of this attribute with the value of the service_id which has
the python type `None` in that case.

If any existing deployments were created without the service_id
attribute, we now have to support this case and make sure the check
won't filter out devices unexpectedly.

Fixes: https://tracker.ceph.com/issues/63729

Signed-off-by: Guillaume Abrioux <gabrioux@ibm.com>
(cherry picked from commit c68b5af0fb639fccc89d26606c7924c6834bf606)