osd: fix Incremental decode for new/old_pg_upmap_primary
Should have been included in commit cc0ad1345bcaa7f8cd8439d785d5c84615f379ac.
The impact is not major; the effect is that the osdmap
is not properly updated whenever the Incremental is reused.
Since the current read balancer offline tool uses a new Incremental
every time it's run, the balancing works as expected. This fix will be
necessary, however, for the mgr module integration in https://github.com/ceph/ceph/pull/53988
since we will need to reuse the Incremental.

Signed-off-by: Laura Flores <lflores@ibm.com>