pybind/mgr/orchestrator: workaround spurious mypy issue
When run locally mypy complains about the `name` field, possibly due to
us using an old mypy version on newer versions of python (we're stuck on
a Sept. 2022 version if you weren't aware). However, this could still be
an issue in mypy as there are open issues related to the `.name` field
of an enum. Regardless, we wouldn't know - so just work around it.

Signed-off-by: John Mulligan <jmulligan@redhat.com>