rgw/multisite: move the archive object instance generator outside the if{} block.
the instance generation was conditional upon versioned_epoch being 0.
this epoch is the source versioned_epoch. but since we require archive zone
to be always versioned, we force versioning even when source is unversioned.
however, if the source is also a versioned bucket, then this versioned_epoch
might not be 0, so we may never enter this if block. Also the instance.key
will not be empty either.

this commit removes the instance.key check and also moves the instance generation
outside the if block so that we consistently create static object versions in all
states i.e enabled, suspended and non versioned source buckets.

Signed-off-by: Shilpa Jagannath <smanjara@redhat.com>