mon/AuthMonitor: fix potential repeated global id
When we expand or shrink monitors, there is the possibility of
repeatedly allocating global ids. Because the calculation of
last_allocated_id depends on mon_num and mon_rank, and expand or
shrink monitors changes mon_num and mon_rank. A simple
fix is to reset last_allocated_id when the election is complete.

Fixs: https://tracker.ceph.com/issues/63891
Signed-off-by: shimin <shimin@kuaishou.com>