win32_deps_build.sh: download Boost from Ceph mirror
We've recently changed the Boost mirror since the previous one
became unavailable.

As pointed out in PR comments, it's even better to use the Ceph
hosted source in order to save Ceph lab bandwidth.

We're adding bzip2 as a dependency. xz-utils was missing, so we're
adding it to our list.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>