json_spirit: Add a long constructor for json_spirit::Value
clang on Darwin has troubles picking a constructor of Value_impl for an implicit
conversion from `long`, for example on MacOS:

src/tools/rbd_mirror/image_replayer/snapshot/Replayer.cc:289:74:
  error: conversion from 'time_t' (aka 'long') to 'const Value_impl<Config_map<string>>' is ambiguous
    root_obj["syncing_snapshot_timestamp"] = remote_snap_info->timestamp.sec();

This appears to happen due to the fact that on MacOS `long` and `int64_t` are
distinct types, even though long is 8 bytes. int64_t is defined as `long long`.

On a 64 bit linux and gcc `long` is apprently equivalent to int64_t, so it has
no trouble selecting the constructor, and even more than that, barks
at a separate `long` constructor as a duplicate. The constraint requirement
is a trick to make everyone happy.

Signed-off-by: Leonid Usov <leonid.usov@ibm.com>