pybind/mgr/dashboard: correct type signature of CRUDEndpoint arg
Typing issue was found by mypy 1.5.1. The type only accepted values
of str but based on the data and code using set_column data the
boolean assignments to the 'isHidden' key were correct and so
I updated the type to also accept bool.

A more correct fix might be to use typing.TypedDict - this would
require python 3.8 (or typing extensions) and someone more familiar
with the dashboard code :-)

Signed-off-by: John Mulligan <jmulligan@redhat.com>