mds/quiesce: QuiesceDb: Update the semantics of the `--reset` command
After discussing this with the CSI team it became clear that we
should allow the --reset command to force a set with the given id
to be moved to QUIESCING state even if it's already in some terminal state.

CSI will operate under a k8s reconcile loop which may result in the same
command being retried after past failures. Using the same command id as the
set identifier is the assumed operation mode.

Fixes: https://tracker.ceph.com/issues/64252
Signed-off-by: Leonid Usov <leonid.usov@ibm.com>