rgw: if head object does not have data, still include in `object manifest`
In order to enable users who might like to use `radosgw-admin object
manifest ...` to get a complete list of raw objects that comprise the
RGW object, we'll include the head object when it includes not
data. It will be indicated with an "index" of -1.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit 575362d826f8f58832a2bb865b9dc73c474f4fe1)