rgw/auth: Refactor Keystone EC2Engine to not use admin token
The EC2Engine::get_from_keystone() function makes a HTTP
POST request to /v3/s3tokens that is a public API [1] that
does not require authentication to use, so we can drop the
admin token in the header in that request.

The EC2Engine::get_secret_from_keystone() function makes
the following HTTP GET request to get the secret key for
that credential using the token we retrieved in the
get_from_keystone function [2].

  /v3/users/{user_id}/credentials/OS-EC2/{credential_id}

This HTTP GET request in Keystone is validated by the
identity:ec2_get_credential policy that before required
an admin token but today can be retrieved by the
identity that created the credential [3].

That means we can drop the usage of admin token in the
Keystone EC2Engine if we bump the minimum Keystone version
required to >=16.0.0 (Train release) [4] that has been
considered Extended Maintenance (EM) for a long time [5]
and has since had it's git branch stable/train EOL'd into
a train-eol git tag.

[1] https://github.com/openstack/keystone/blob/master/keystone/api/s3tokens.py#L94
[2] https://github.com/openstack/keystone/blob/master/keystone/api/users.py#L420
[3] https://bugs.launchpad.net/keystone/+bug/1750678
[4] https://releases.openstack.org/train/index.html
[5] https://review.opendev.org/c/openstack/releases/+/790750

Signed-off-by: Tobias Urdin <tobias.urdin@binero.se>