multi-mds: Make sure destdn dirfrags are loaded
The is done to fix the race between hardlink creation (mds.x)
and lookup (mds.y) on auth of the real inode. As part of
hardlink creation, referent inode is created at 'mds.x' and
added to the referent inode list in 'mds.y' as part 'link_prep'
peer op. After the hard link peer is commited at 'mds.y', the
getattr on primary file comes which could not to load the referent
inode.

The loading of dirfrags fixed the issue. But after looking into
the logs, it's found out that getattr initiated a path_traverse
on the primary file's inode (no tail dentry) and hence the referent
is not loaded. This part wasn't handled.

TODO :
Now that's been in handled in 'path_traverse', do we need really
need to load the destdn dirfrags during hard link creation ?
I think no, this needs to be validated.

Signed-off-by: Kotresh HR <khiremat@redhat.com>