qa/workunits: opt in to new collection after upgrade from reef to main
We added the "basic_pool_flags" collection to main (squid), so
we need to re-opt-in after the upgrade to opt in to this collection.

Otherwise, a "TELEMETRY_CHANGED" warning comes up in the log and fails
the test.

Fixes: https://tracker.ceph.com/issues/64458
Signed-off-by: Laura Flores <lflores@ibm.com>
(cherry picked from commit a04a9c64b86fc253997081c4d769c887d13166a1)