common: improve hobject_t fmtlib formatter performance
The new version of the code only takes about 38% of the time of the old
one.  See https://github.com/ronen-fr/hobjtostr/tree/rf-2 for the code
used to benchmark the new version.

hobject_fmt.h is folded into hobject.h, as fmtlib is now an accepted
dependency in all of Ceph.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>